Onboarding improvements + build fix (#41)
The functional effect of this PR is to add a `docker_desktop` query
param (which doesn't yet do anything) to the extension's link for
creating an Akita account, and to move "Project" above the api key and
api key secret fields in the frontend. This ordering change corresponds
to the order the project and key are created during the onboarding in
the Akita app.

Also, #40 left a couple of lint errors that block the `make` command for
building the extension from working; this happened because the extension
was already built, and the changes that made it into the diff were
hot-reloaded.